x64 version (#1113)
* Sort out compiler errors (excluding external libraries)

* Add TODO comment

* Use empty() method

* Fix wrong casts

* Fix new warnings

* Fix merge error

* Revert "Merge branch 'develop' into sezz_x64"

This reverts commit f6957691890a8f6fbbd9ef5d83ad6fbe8e174193, reversing
changes made to 54c5e0c70d0f88c92c74dba368f12e1715561437.

* Revert "Revert "Merge branch 'develop' into sezz_x64""

This reverts commit e1128c41f87f2ace92f710cacc1c8427f4acae37.

* Update all libraries with x86 versions, organize lib directories, update x64 config

* Show app bitness in a log file

* Fix text rendering by uncommenting ToWString helper function

* Ship dlls and automatically replace them when switching between x86/x64

* Update TombEngine.vcxproj

* Adjust ammo struct; remove unneeded line

* Update broken x64 config for release mode

* Fix more project config inconsistencies

* Update TombEngine.vcxproj

* Remove unnecessary casts

* Tabs not spaces

* Update TombEngine.vcxproj.user

* Revert "Update TombEngine.vcxproj.user"

This reverts commit c168943ed0cb855eacda889df7a06148e59cb692.

* Add x64 lua53.lib, remove DLLs.

---------

+++ TombEngine/Scripting/Internal/TEN/Flow/FlowHandler.cpp
@@ -313,7 +313,7 @@ Level* FlowHandler::GetCurrentLevel()
 
 int	FlowHandler::GetNumLevels() const
 {
-	return Levels.size();
+	return (int)Levels.size();
 }
 
 int FlowHandler::GetLevelNumber(const std::string& fileName)
